Split out `error.hh` to its own file
